requirements.txt
openai>=1.0
google-genai>=0.1
anthropic>=0.20
Pillow>=10.0
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
